third_party/blink/web_tests/http/tests/devtools/profiler/cpu-profiler-stopped-removed-race-expected.txt

Tests that profile removal right after stop profiling issued works. Bug 476430.

Profiler was enabled.

Running: testProfiling
Suspending targets
Stop recording command issued.
Recording stopped. profile (should be null): null
Resuming targets

Profiler was disabled.
